feat(TDI-41956) : some cleaning
