+++ common/src/main/java/org/talend/components/common/service/http/common/BasicHeaderValueParser.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.talend.components.common.service.http;
+package org.talend.components.common.service.http.common;
 
 import java.util.Arrays;
 import java.util.HashMap;
