+++ common/src/main/java/org/talend/components/common/service/http/common/ParserCursor.java
@@ -10,9 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.talend.components.common.service.http;
-
-import org.talend.components.common.service.http.CharArrayBuffer;
+package org.talend.components.common.service.http.common;
 
 /**
  * This class represents a context of a parsing operation:
